requirements.txt
python-telegram-bot
selenium
pyyaml
marshmallow
argparse
discord
ruamel